backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
pyodbc==5.0.1
mssql-django==1.3
django-cors-headers==4.3.0

